fix: validate celldm(1) and integer nbnd at parse time
celldms without key 1 previously surfaced as a bare KeyError deep in the converter; nbnd was typed float and silently truncated fractional values.

+++ src/koopmans/input_file/__init__.py
@@ -132,7 +132,7 @@ class CalculatorParametersInput(BaseModel):
     """Calculator-specific input parameters."""
 
     ecutwfc: float | None = None
-    nbnd: float | None = None
+    nbnd: int | None = None
     tot_magnetization: float | None = None
     pw: PWInputParameters = Field(default_factory=lambda: PWInputParameters())
     pw2wannier90: PW2Wannier90InputParameters = Field(
